[config] renamed from _dev_config
